[BUG] - Awake not adding listener
