+++ Assets/_Scenes/NewBattle.cs.unity
@@ -9593,6 +9593,7 @@ GameObject:
   - component: {fileID: 2026179254}
   - component: {fileID: 2026179255}
   - component: {fileID: 2026179244}
+  - component: {fileID: 2026179256}
   m_Layer: 0
   m_Name: StageController
   m_TagString: Untagged
@@ -9812,6 +9813,18 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   callback: NEXT_PHASE
+--- !u!114 &2026179256
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2026179242}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 7196084fea5474497b52d9d95fcc498e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &2029645778
 GameObject:
   m_ObjectHideFlags: 0
