+++ ProjectSettings/EditorBuildSettings.asset
@@ -21,6 +21,6 @@ EditorBuildSettings:
     path: Assets/_Scenes/WorldMap.unity
     guid: 40e3d7292c2c09e48be0214605198011
   - enabled: 1
-    path: Assets/_Scenes/Battle.unity
-    guid: 8bb1843e6db80ae4eb4dda2b45316eb2
+    path: Assets/_Scenes/NewBattle.cs.unity
+    guid: 541ef4bedc1b549db9c462cffa190d94
   m_configObjects: {}
